Remoção dos templates, passando para o pytrustnfe e mudando os metodos para chamar a lib
