+++ nfse_campinas/models/account_invoice.py
@@ -40,23 +40,6 @@ class AccountInvoice(models.Model):
     def _default_taxation(self):
         return self.env.user.company_id.default_taxation
 
-    def _default_state(self):
-        if self.env.user.company_id.state_id:
-            return self.env.user.company_id.state_id
-
-    def _default_city(self):
-        if self.env.user.company_id.l10n_br_city_id:
-            return self.env.user.company_id.l10n_br_city_id
-
-    state_id = fields.Many2one('res.country.state', string=u"Estado",
-                               default=_default_state,
-                               domain=[('country_id.code', '=', 'BR')])
-
-    provider_city_id = fields.Many2one('l10n_br_base.city',
-                                       string=u"Munícipio Prestação",
-                                       readonly=True, default=_default_city,
-                                       states=FIELD_STATE)
-
     type_retention = fields.Selection([('A', u'ISS a recolher pelo prestador'),
                                        ('R', u'Retido na Fonte')],
                                       u'Tipo Recolhimento',
