--- nfse_sao_paulo/templates/consulta_cnpj.xml
@@ -1,12 +1,0 @@
-<p1:PedidoConsultaCNPJ 
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
-  xmlns:p1="http://www.prefeitura.sp.gov.br/nfe">
-  <Cabecalho Versao="1">
-    <CPFCNPJRemetente>
-      <CNPJ>{{ nfse.cpf_cnpj }}</CNPJ>
-    </CPFCNPJRemetente>
-  </Cabecalho>
-  <CNPJContribuinte>
-      <CNPJ>08944335000170</CNPJ>
-  </CNPJContribuinte>
-</p1:PedidoConsultaCNPJ>
