--- nfse_sao_paulo/templates/consulta_lote.xml
@@ -1,10 +1,0 @@
-<ns1:ReqConsultaLote xmlns:ns1="http://localhost:8080/WsNFe2/lote"
-	xmlns:tipos="http://localhost:8080/WsNFe2/tp" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://localhost:8080/WsNFe2/lote http://localhost:8080/WsNFe2/xsd/ReqConsultaLote.xsd">
-	<Cabecalho>
-		<CodCidade>{{ consulta.cidade }}</CodCidade>
-		<CPFCNPJRemetente>{{ consulta.cpf_cnpj }}</CPFCNPJRemetente>
-		<Versao>1</Versao>
-		<NumeroLote>{{ consulta.lote }}</NumeroLote>
-	</Cabecalho>
-</ns1:ReqConsultaLote>
