--- nfse_sao_paulo/templates/envio_loterps.xml
@@ -1,54 +1,0 @@
-<PedidoEnvioLoteRPS 
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
-  xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns="http://www.prefeitura.sp.gov.br/nfe">
-  <Cabecalho Versao="1" xmlns="">
-    <CPFCNPJRemetente>
-      <CNPJ>{{ nfse.cpf_cnpj }}</CNPJ>
-    </CPFCNPJRemetente>
-    <transacao>false</transacao>
-    <dtInicio>2016-02-08</dtInicio>
-    <dtFim>2016-02-08</dtFim>
-    <QtdRPS>1</QtdRPS>
-    <ValorTotalServicos>{{ nfse.total_servicos }}</ValorTotalServicos>
-    <ValorTotalDeducoes>{{ nfse.total_deducoes }}</ValorTotalDeducoes>
-  </Cabecalho>
-  {% for rps in nfse.lista_rps -%}
-  <RPS xmlns="">
-    <Assinatura>{{ rps.assinatura }}</Assinatura>
-    <ChaveRPS>
-      <InscricaoPrestador>{{ rps.prestador.inscricao_municipal }}</InscricaoPrestador>
-      <SerieRPS>{{ rps.serie }}</SerieRPS>
-      <NumeroRPS>{{ rps.numero }}</NumeroRPS>
-    </ChaveRPS>
-    <TipoRPS>RPS</TipoRPS>
-    <DataEmissao>{{ rps.data_emissao }}</DataEmissao>
-    <StatusRPS>N</StatusRPS>
-    <TributacaoRPS>T</TributacaoRPS>
-    <ValorServicos>1000</ValorServicos>
-    <ValorDeducoes>100</ValorDeducoes>
-    <CodigoServico>{{ rps.codigo_atividade }}</CodigoServico>
-    <AliquotaServicos>{{ rps.aliquota_atividade }}</AliquotaServicos>
-    <ISSRetido>false</ISSRetido>
-    <CPFCNPJTomador>
-        {% if rps.tomador.tipo_cpfcnpj == 1 -%}
-           <CPF>{{ rps.tomador.cpf_cnpj }}</CPF>
-        {% endif %}
-        {% if rps.tomador.tipo_cpfcnpj == 2 -%}
-           <CNPJ>{{ rps.tomador.cpf_cnpj }}</CNPJ>
-        {% endif %}
-    </CPFCNPJTomador>
-    <RazaoSocialTomador>{{ rps.tomador.razao_social }}</RazaoSocialTomador>
-    <EnderecoTomador>
-      <TipoLogradouro>{{ rps.tomador.tipo_logradouro }}</TipoLogradouro>
-      <Logradouro>{{ rps.tomador.logradouro }}</Logradouro>
-      <NumeroEndereco>{{ rps.tomador.numero }}</NumeroEndereco>
-      <ComplementoEndereco>{{ rps.tomador.complemento }}</ComplementoEndereco>
-      <Bairro>{{ rps.tomador.bairro }}</Bairro>
-      <Cidade>{{ rps.tomador.cidade }}</Cidade>
-      <UF>{{ rps.tomador.uf }}</UF>
-      <CEP>{{ rps.tomador.cep }}</CEP>
-    </EnderecoTomador>
-    <Discriminacao>{{ rps.descricao }}</Discriminacao>
-  {% endfor %}
-  </RPS>
-</PedidoEnvioLoteRPS>
