--- nfse_sao_paulo/templates/soap_header.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<soapenv:Envelope xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/"
-	xmlns:dsf="http://dsfnet.com.br">
-	<soapenv:Body>
-		<dsf:enviar soapenv:encodingStyle="http://schemas.xmlsoap.org/soap/encoding/">
-			<mensagemXml xsi:type="xsd:string"><![CDATA[
-				{% block content %}{% endblock %}
-			]]></mensagemXml>
-		</dsf:enviar>
-	</soapenv:Body>
-</soapenv:Envelope>
